homework done.

+++ src/homework.java
@@ -2,7 +2,7 @@ public class homework {
     public static void main(String[] args) {
         char b = 'A';
         int c;
-        c = (int) b;
+        c = b;
         System.out.println(c);
